# shared/contracts/embedding-api.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Guised Up Embedding Service",
    "version": "0.1.0",
    "description": "Internal FastAPI embedding inference service (TSD §24). Not exposed publicly."
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Service healthy and model loaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "model", "version"],
                  "properties": {
                    "status": { "type": "string", "example": "ok" },
                    "model": { "type": "string", "example": "all-MiniLM-L6-v2" },
                    "version": { "type": "string", "example": "3.3.1" }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Model not ready"
          }
        }
      }
    },
    "/embed": {
      "post": {
        "summary": "Generate embedding for a single text",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": {
                  "text": { "type": "string", "minLength": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Embedding generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["embedding", "model"],
                  "properties": {
                    "embedding": {
                      "type": "array",
                      "items": { "type": "number" },
                      "minItems": 384,
                      "maxItems": 384
                    },
                    "model": { "type": "string" }
                  }
                }
              }
            }
          },
          "422": { "description": "Validation error" },
          "504": { "description": "Inference timeout" }
        }
      }
    },
    "/embed/batch": {
      "post": {
        "summary": "Generate embeddings for multiple texts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["texts"],
                "properties": {
                  "texts": {
                    "type": "array",
                    "minItems": 1,
                    "items": { "type": "string", "minLength": 1 }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Embeddings generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["embeddings", "model"],
                  "properties": {
                    "embeddings": {
                      "type": "array",
                      "items": {
                        "type": "array",
                        "items": { "type": "number" },
                        "minItems": 384,
                        "maxItems": 384
                      }
                    },
                    "model": { "type": "string" }
                  }
                }
              }
            }
          },
          "422": { "description": "Validation error" },
          "504": { "description": "Inference timeout" }
        }
      }
    }
  }
}